refactor: move shared route test helpers
