+++ backend/tests/routes/support/__init__.py
@@ -1,0 +1,13 @@
+"""Shared route test support helpers"""
+
+from tests.routes.support.account_helpers import ACCOUNT_PAYLOAD, _create_account
+from tests.routes.support.auth_helpers import SIGNUP_PAYLOAD, _create_user, _get_auth_header, _seed_currency
+
+__all__ = [
+    "ACCOUNT_PAYLOAD",
+    "SIGNUP_PAYLOAD",
+    "_create_account",
+    "_create_user",
+    "_get_auth_header",
+    "_seed_currency",
+]
